@@ -30,8 +30,10 @@ dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^0.1.2
-  charts_flutter: ^0.9.0
-  wakelock: ^0.1.4+1
+  charts_flutter: ^0.12.0
+  charts_common: ^0.12.0
+  intl: ^0.17.0
+  wakelock: ^0.6.1+1
   camera:
   #   git:
   #     url: https://github.com/Renesanse/plugins.git
